Move sassdocs_helpers into lib folder
To match Ruby conventions as we're going to have more than one ruby file in our project
