+++ lib/sassdocs_helpers.rb
